requirements.txt
Django==3.1.5
django-crispy-forms==1.10.0
django-cleanup==5.1.0
Pillow==8.1.0
psycopg2==2.8.6
django-heroku==0.3.1
whitenoise==5.2.0
gunicorn==20.0.4
phonenumbers==8.12.37
django-phonenumber-field
django-location-field